Refactiored to have  handler outside Header, added Teleport dialog opening as well
